release: prepare for release 2.2.0

+++ platforms/ble/battery_driver.go
@@ -57,6 +57,6 @@ func (b *BatteryDriver) GetBatteryLevel() (level uint8) {
 	}
 	buf := bytes.NewBuffer(c)
 	val, _ := buf.ReadByte()
-	l = uint8(val)
+	l = val
 	return l
 }
